// apps/raise/src/tender-statistics/dtos/responses/get-average-transaction.dto.ts
export class GetTrackAverageTransaction {
  project_track: string;
  raw_total_response_time: number;
  total_response_time: string;
  raw_average_response_time: number;
  average_response_time: string;
  raw_last_month_total_response_time: number;
  last_month_total_response_time: string;
  raw_last_month_average_response_time: number;
  last_month_average_response_time: string;
}