// apps/raise/src/tender-statistics/services/tender-statistics.service.ts
import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
import * as _ from 'lodash';
import moment from 'moment';
import { logUtil } from '../../commons/utils/log-util';
import { PrismaService } from '../../prisma/prisma.service';
import { portalReportConvertMinutesToHours } from '../../tender-commons/utils/portal-report-convert-minutes-to-hour';
import { AverageEmployeeTransactionTimeFilter } from '../dtos/requests/average-employee-transaction-time-filter';
import { BaseStatisticFilter } from '../dtos/requests/base-statistic-filter.dto';
import {
  AverageEmployeeResponseTime,
  GetAverageEmployeeResponseTime,
} from '../dtos/responses/get-average-employee-response-time.dto';
import { GetTrackAverageTransaction } from '../dtos/responses/get-average-transaction.dto';
import {
  GetBeneficiariesReportDto,
  IGetBeneficiariesByTrackDto,
  IGetBeneficiariesByTypeDto,
} from '../dtos/responses/get-beneficiaries-report.dto';
import {
  GetPartnersStatisticByGovernorateResponseDto,
  GetPartnersStatisticByRegionResponseDto,
  GetPartnersStatisticResponseDto,
  PartnerValue,
} from '../dtos/responses/get-partner-statistic.dto';
import { GetRawAverageEmployeeResponseTime } from '../dtos/responses/get-raw-average-employee-response-time.dto';
import { TenderStatisticsRepository } from '../repositories/tender-statistic.repository';

@Injectable()
export class TenderStatisticsService {
  constructor(
    private readonly prismaService: PrismaService,
    private readonly tenderStatisticRepository: TenderStatisticsRepository,
  ) {}

  async getAllStatistics(from: any, to: any) {
    const tarcksStatistics = await this.getAllTrackStatistics(from, to);
    return tarcksStatistics;
  }

  async getAllTrackStatistics(from: any, to: any) {
    const response = {} as any;
    const governorates = await this.prismaService.proposal.findMany({
      where: {
        AND: [
          {
            created_at: {
              gte: from,
            },
          },
          {
            created_at: {
              lte: to,
            },
          },
        ],
      },
      select: {
        governorate: true,
        outter_status: true,
        project_track: true,
      },
    });
    const governorateObj = {} as any;
    const nestedGovernorateObj = {} as any;
    const trackGovernorateObj = {} as any;
    governorates.map((governorate) => {
      if (
        !governorate.governorate ||
        !governorate.outter_status ||
        !governorate.project_track
      )
        return;
      if (!governorateObj[governorate.governorate]) {
        nestedGovernorateObj[governorate.outter_status] = 1;
        trackGovernorateObj[governorate.project_track] = Object.assign(
          {},
          nestedGovernorateObj,
        );
        governorateObj[governorate.governorate] = Object.assign(
          {},
          trackGovernorateObj,
        );
        delete nestedGovernorateObj[governorate.outter_status];
        delete trackGovernorateObj[governorate.project_track];
      } else {
        if (
          !governorateObj[governorate.governorate][governorate.project_track]
        ) {
          nestedGovernorateObj[governorate.outter_status] = 1;
          trackGovernorateObj[governorate.project_track] = Object.assign(
            {},
            nestedGovernorateObj,
          );
          governorateObj[governorate.governorate] = Object.assign(
            governorateObj[governorate.governorate],
            trackGovernorateObj,
          );
          delete nestedGovernorateObj[governorate.outter_status];
          delete trackGovernorateObj[governorate.project_track];
        } else {
          if (
            !governorateObj[governorate.governorate][governorate.project_track][
              governorate.outter_status
            ]
          ) {
            nestedGovernorateObj[governorate.outter_status] = 1;
            governorateObj[governorate.governorate][governorate.project_track] =
              Object.assign(
                governorateObj[governorate.governorate][
                  governorate.project_track
                ],
                nestedGovernorateObj,
              );
            delete nestedGovernorateObj[governorate.outter_status];
          } else {
            governorateObj[governorate.governorate][governorate.project_track][
              governorate.outter_status
            ] =
              governorateObj[governorate.governorate][
                governorate.project_track
              ][governorate.outter_status] + 1;
          }
        }
      }
    });
    response.governorates = governorateObj;
    const regions = await this.prismaService.proposal.findMany({
      where: {
        AND: [
          {
            created_at: {
              gte: from,
            },
          },
          {
            created_at: {
              lte: to,
            },
          },
        ],
      },
      select: {
        region: true,
        outter_status: true,
        project_track: true,
      },
    });
    const regionObj = {} as any;
    const nestedRegionObj = {} as any;
    const trackRegionObj = {} as any;
    regions.map((region) => {
      if (!region.region || !region.outter_status || !region.project_track)
        return;
      if (!regionObj[region.region]) {
        nestedRegionObj[region.outter_status] = 1;
        trackRegionObj[region.project_track] = Object.assign(
          {},
          nestedRegionObj,
        );
        regionObj[region.region] = Object.assign({}, trackRegionObj);
        delete nestedRegionObj[region.outter_status];
        delete trackRegionObj[region.project_track];
      } else {
        if (!regionObj[region.region][region.project_track]) {
          nestedRegionObj[region.outter_status] = 1;
          trackRegionObj[region.project_track] = Object.assign(
            {},
            nestedRegionObj,
          );
          regionObj[region.region] = Object.assign(
            regionObj[region.region],
            trackRegionObj,
          );
          delete nestedRegionObj[region.outter_status];
          delete trackRegionObj[region.project_track];
        } else {
          if (
            !regionObj[region.region][region.project_track][
              region.outter_status
            ]
          ) {
            nestedRegionObj[region.outter_status] = 1;
            regionObj[region.region][region.project_track] = Object.assign(
              regionObj[region.region][region.project_track],
              nestedRegionObj,
            );
            delete nestedRegionObj[region.outter_status];
          } else {
            regionObj[region.region][region.project_track][
              region.outter_status
            ] =
              regionObj[region.region][region.project_track][
                region.outter_status
              ] + 1;
          }
        }
      }
    });
    response.regions = regionObj;
    const authorities = await this.prismaService.client_data.groupBy({
      where: {
        AND: [
          {
            created_at: {
              gte: from,
            },
          },
          {
            created_at: {
              lte: to,
            },
          },
        ],
      },
      by: ['authority', 'user_id'],
    });
    const obj = {} as any;
    authorities.map((auth) => {
      if (!auth.authority) return;
      if (!obj[auth.authority]) obj[auth.authority] = [auth.user_id];
      else obj[auth.authority].push(auth.user_id);
    });
    const authorityObj = {} as any;
    const nestedAuthorityObj = {} as any;
    for (const [authority, userIds] of Object.entries(obj)) {
      const authorities = await this.prismaService.proposal.findMany({
        where: {
          AND: [
            {
              created_at: {
                gte: from,
              },
            },
            {
              created_at: {
                lte: to,
              },
            },
            {
              submitter_user_id: { in: userIds as any },
            },
          ],
        },
        select: {
          outter_status: true,
          project_track: true,
        },
      });
      authorities.map((authority) => {
        if (!authority.outter_status || !authority.project_track) return;
        if (!authorityObj[authority.project_track]) {
          nestedAuthorityObj[authority.outter_status] = 1;
          authorityObj[authority.project_track] = Object.assign(
            {},
            nestedAuthorityObj,
          );
          delete nestedAuthorityObj[authority.outter_status];
        } else {
          if (!authorityObj[authority.project_track][authority.outter_status]) {
            nestedAuthorityObj[authority.outter_status] = 1;
            authorityObj[authority.project_track] = Object.assign(
              authorityObj[authority.project_track],
              nestedAuthorityObj,
            );
            delete nestedAuthorityObj[authority.outter_status];
          } else {
            authorityObj[authority.project_track][authority.outter_status] =
              authorityObj[authority.project_track][authority.outter_status] +
              1;
          }
        }
      });
    }
    response.authorities = authorityObj;

    const tracksObj = {} as any;
    const tracks = await this.prismaService.proposal.findMany({
      where: {
        AND: [
          {
            created_at: {
              gte: from,
            },
          },
          {
            created_at: {
              lte: to,
            },
          },
        ],
      },
      select: {
        outter_status: true,
        project_track: true,
      },
    });
    const nestedTrackObj = {} as any;
    tracks.map((track) => {
      if (!track.project_track || !track.outter_status) return;
      if (!tracksObj[track.project_track]) {
        nestedTrackObj[track.outter_status] = 1;
        tracksObj[track.project_track] = Object.assign({}, nestedTrackObj);
        delete nestedTrackObj[track.outter_status];
      } else {
        if (!tracksObj[track.project_track][track.outter_status]) {
          nestedTrackObj[track.outter_status] = 1;
          tracksObj[track.project_track] = Object.assign(
            tracksObj[track.project_track],
            nestedTrackObj,
          );
          delete nestedTrackObj[track.outter_status];
        } else {
          tracksObj[track.project_track][track.outter_status] =
            tracksObj[track.project_track][track.outter_status] + 1;
        }
      }
    });
    response.tracks = tracksObj;
    return response;
  }

  async getAllParntersStatistics(from: any, to: any) {
    const response = {} as any;
    const userStatus = await this.prismaService.user_status.findMany();
    const partnersStatus = {} as any;
    const partnersRegion = {} as any;
    const nestedPartnersRegion = {} as any;
    const partnersGovernorate = {} as any;
    const nestedPartnersGovernorate = {} as any;

    for (const US of userStatus) {
      const pendingUsers = await this.prismaService.client_data.findMany({
        where: {
          AND: [
            {
              created_at: {
                gte: from,
              },
            },
            {
              created_at: {
                lte: to,
              },
            },
          ],
        },
        include: {
          user: {
            select: {
              id: true,
              status_id: true,
            },
          },
        },
      });

      for (const PU of pendingUsers) {
        if (!partnersStatus[US.id]) partnersStatus[US.id] = 0;
        if (PU.user.status_id === US.id) {
          partnersStatus[US.id] = partnersStatus[US.id] + 1;
        }

        if (!PU.region) return;
        if (!partnersRegion[PU.region]) {
          nestedPartnersRegion[US.id] = 1;
          partnersRegion[PU.region] = Object.assign({}, nestedPartnersRegion);
          delete nestedPartnersRegion[US.id];
        } else {
          if (!partnersRegion[PU.region][US.id]) {
            nestedPartnersRegion[US.id] = 1;
            partnersRegion[PU.region] = Object.assign({}, nestedPartnersRegion);
            delete nestedPartnersRegion[US.id];
          } else {
            partnersRegion[PU.region][US.id] =
              partnersRegion[PU.region][US.id] + 1;
          }
        }

        if (!PU.governorate) return;
        if (!partnersGovernorate[PU.governorate]) {
          nestedPartnersGovernorate[US.id] = 1;
          partnersGovernorate[PU.governorate] = Object.assign(
            {},
            nestedPartnersGovernorate,
          );
          delete nestedPartnersGovernorate[US.id];
        } else {
          if (!partnersGovernorate[PU.governorate][US.id]) {
            nestedPartnersGovernorate[US.id] = 1;
            partnersGovernorate[PU.governorate] = Object.assign(
              {},
              nestedPartnersGovernorate,
            );
            delete nestedPartnersGovernorate[US.id];
          } else {
            partnersGovernorate[PU.governorate][US.id] =
              partnersGovernorate[PU.governorate][US.id] + 1;
          }
        }
      }
    }
    response.partnersStatus = partnersStatus;
    response.partnersRegion = partnersRegion;
    response.partnersGovernorate = partnersGovernorate;

    return response;
  }

  async getPartnersStatistic(
    filter: BaseStatisticFilter,
  ): Promise<GetPartnersStatisticResponseDto> {
    const partners = await this.tenderStatisticRepository.getRawPartnersData(
      filter,
    );

    const partnerStatus =
      await this.tenderStatisticRepository.getRawUserStatus();

    if (partners.length > 0) {
      const latestPartnerData = moment(filter.end_date).startOf('day').toDate();
      const byStatus = partnerStatus.map((status) => {
        return {
          label: status.id,
          value: partners.filter(
            (partner) => partner.user_status.id === status.id,
          ).length,
        };
      });

      // create Maps for byRegion and byGovernorate
      const byRegion = new Map<string, any>();
      const byGovernorate = new Map<string, any>();

      partners.forEach((partner) => {
        const { user_detail } = partner;
        const { client_data } = user_detail;
        if (!client_data) return;
        const { region, governorate } = client_data;
        const { id: userId } = user_detail;
        const status = partner.user_status.id;

        // add to byRegion map
        if (region) {
          if (byRegion.has(region)) {
            const currentData: GetPartnersStatisticByRegionResponseDto =
              byRegion.get(region);
            if (
              !currentData.value.find(
                (val: PartnerValue) => val.user_id === userId,
              )
            ) {
              currentData.value.push({ user_id: userId, label: [status] });
              currentData.total++;
            } else {
              currentData.value
                .find((val: PartnerValue) => val.user_id === userId)!
                .label.push(status);
            }
          } else {
            byRegion.set(region, {
              label: region,
              value: [{ user_id: userId, label: [status] }],
              total: 1,
            });
          }
        }

        // add to byGovernorate map
        if (governorate) {
          if (byGovernorate.has(governorate)) {
            const currentData: GetPartnersStatisticByGovernorateResponseDto =
              byGovernorate.get(governorate);
            if (
              !currentData.value.find(
                (val: PartnerValue) => val.user_id === userId,
              )
            ) {
              currentData.value.push({ user_id: userId, label: [status] });
              currentData.total++;
            } else {
              currentData.value
                .find((val: PartnerValue) => val.user_id === userId)!
                .label.push(status);
            }
          } else {
            byGovernorate.set(governorate, {
              label: governorate,
              value: [{ user_id: userId, label: [status] }],
              total: 1,
            });
          }
        }
      });

      //calculate monthly data
      const monthlyData = {
        this_month: partnerStatus.map((status) => {
          return {
            label: status.id,
            value: partners.filter(
              (partner) =>
                partner.user_status.id === status.id &&
                partner.created_at!.getMonth() === latestPartnerData.getMonth(),
            ).length,
          };
        }),
        last_month: partnerStatus.map((status) => {
          return {
            label: status.id,
            value: partners.filter(
              (partner) =>
                partner.user_status.id === status.id &&
                partner.created_at!.getMonth() ===
                  latestPartnerData.getMonth() - 1,
            ).length,
          };
        }),
      };
      return {
        by_status: byStatus,
        by_region: Array.from(byRegion.values()),
        by_governorate: Array.from(byGovernorate.values()),
        monthlyData: monthlyData,
      };
    } else {
      return {
        by_status: [],
        by_region: [],
        by_governorate: [],
        monthlyData: {
          this_month: [],
          last_month: [],
        },
      };
    }
  }

  async getBeneficiariesReport(
    filter: BaseStatisticFilter,
  ): Promise<GetBeneficiariesReportDto> {
    // Get raw data from repository (filtering will be done in the repository)
    const proposals =
      await this.tenderStatisticRepository.getRawBeneficiariesData(filter);

    // Initialize empty objects for by_track and by_type data
    const byTrack: { [key: string]: IGetBeneficiariesByTrackDto } = {};
    const byType: { [key: string]: IGetBeneficiariesByTypeDto } = {};

    if (proposals.length > 0) {
      // Loop through raw data
      for (let i = 0; i < proposals.length; i++) {
        const proposal = proposals[i];
        // If project_track and num_ofproject_binicficiaries are not null, add data to byTrack object
        if (
          proposal.project_track != null &&
          proposal.num_ofproject_binicficiaries != null
        ) {
          // Check if project_track exists in byTrack object
          if (!byTrack[proposal.project_track]) {
            // If project_track does not exist in byTrack object, initialize it with track and total_project_beneficiaries values
            byTrack[proposal.project_track] = {
              track: proposal.project_track,
              total_project_beneficiaries: 0, // if project_track does not exist, initialize total_project_beneficiaries with 0
            };
          }
          // Add num_ofproject_binicficiaries value to total_project_beneficiaries for the current project_track
          byTrack[proposal.project_track].total_project_beneficiaries +=
            proposal.num_ofproject_binicficiaries;
        }

        // basicly same logic as above
        if (
          proposal.project_beneficiaries != null &&
          proposal.num_ofproject_binicficiaries != null
        ) {
          if (!byType[proposal.project_beneficiaries]) {
            byType[proposal.project_beneficiaries] = {
              type: proposal.project_beneficiaries,
              total_project_beneficiaries: 0,
            };
          }
          byType[proposal.project_beneficiaries].total_project_beneficiaries +=
            proposal.num_ofproject_binicficiaries;
        }
      }

      // Return response with byTrack and byType data
      return {
        by_track: Object.values(byTrack),
        by_type: Object.values(byType),
      };
    } else {
      return {
        by_track: [],
        by_type: [],
      };
    }
  }

  async getBudgetInfo(request: BaseStatisticFilter) {
    const proposals = await this.tenderStatisticRepository.getBudgetInfo(
      request,
    );

    type GroupedProposal = {
      project_track: string | null;
      total_budget: number;
      spended_budget: number;
      spended_budget_last_week: number;
      reserved_budget: number;
      reserved_budget_last_week: number;
    };

    const groupedData: GroupedProposal[] = proposals.reduce(
      (previousData, proposal) => {
        const existingGroup = previousData.find(
          (group) => group.project_track === proposal.project_track,
        );
        if (existingGroup) {
          existingGroup.total_budget += Number(
            proposal.amount_required_fsupport,
          );
          existingGroup.spended_budget += proposal.proposal_item_budgets.reduce(
            (sum, item) => sum + Number(item.amount),
            0,
          );
          // get last week spended budget if proposal.proposal_item_budget.created_at is in last week
          // then add it to existingGroup.spended_budget_last_week if not then add 0
          existingGroup.spended_budget_last_week +=
            proposal.proposal_item_budgets.reduce((sum, item) => {
              // get the last week from the request.end_date if exist,
              // if not get from request.start,
              // if both not exist, get from date now
              let lastWeek: Date;
              if (request.end_date && request.start_date) {
                lastWeek = new Date(request.end_date);
              } else if (request.start_date && !request.end_date) {
                lastWeek = new Date(request.start_date);
              } else {
                lastWeek = new Date();
              }
              lastWeek.setDate(lastWeek.getDate() - 7);
              const date = new Date(item.created_at!);
              if (date > lastWeek) {
                return sum + Number(item.amount);
              }
              return sum;
            }, 0) || 0;
          existingGroup.reserved_budget =
            existingGroup.total_budget - existingGroup.spended_budget;
          existingGroup.reserved_budget_last_week =
            existingGroup.total_budget - existingGroup.spended_budget_last_week;
          return previousData;
        } else {
          previousData.push({
            project_track: proposal.project_track,
            total_budget: Number(proposal.amount_required_fsupport),
            spended_budget: proposal.proposal_item_budgets.reduce(
              (sum, item) => sum + Number(item.amount),
              0,
            ),
            spended_budget_last_week: proposal.proposal_item_budgets.reduce(
              (sum, item) => {
                //   const date = new Date(item.created_at!);
                //   const lastWeek = new Date();
                //   lastWeek.setDate(lastWeek.getDate() - 7);
                //   if (date > lastWeek) {
                //     return sum + Number(item.amount);
                //   }
                //   return sum;
                // },
                // 0,

                // get the last week from the request.end_date if exist,
                // if not get from request.start,
                // if both not exist, get from date now
                let lastWeek: Date;
                if (request.end_date && request.start_date) {
                  lastWeek = new Date(request.end_date);
                } else if (request.start_date && !request.end_date) {
                  lastWeek = new Date(request.start_date);
                } else {
                  lastWeek = new Date();
                }
                lastWeek.setDate(lastWeek.getDate() - 7);
                const date = new Date(item.created_at!);
                if (date > lastWeek) {
                  return sum + Number(item.amount);
                }
                return sum;
              },
              0,
            ),
            reserved_budget:
              0 -
              proposal.proposal_item_budgets.reduce(
                (sum, item) => sum + Number(item.amount),
                0,
              ),
            reserved_budget_last_week:
              0 -
              proposal.proposal_item_budgets.reduce((sum, item) => {
                // const date = new Date(item.created_at!);
                // const lastWeek = new Date();
                // lastWeek.setDate(lastWeek.getDate() - 7);
                // if (date > lastWeek) {
                //   return sum + Number(item.amount);
                // }
                // return sum;

                // get the last week from the request.end_date if exist,
                // if not get from request.start,
                // if both not exist, get from date now
                let lastWeek: Date;
                if (request.end_date && request.start_date) {
                  lastWeek = new Date(request.end_date);
                } else if (request.start_date && !request.end_date) {
                  lastWeek = new Date(request.start_date);
                } else {
                  lastWeek = new Date();
                }
                lastWeek.setDate(lastWeek.getDate() - 7);
                const date = new Date(item.created_at!);
                if (date > lastWeek) {
                  return sum + Number(item.amount);
                }
                return sum;
              }, 0),
          });
        }
        return previousData;
      },
      [] as GroupedProposal[],
    );

    return groupedData;
  }

  async getTrackAverageTransaction(request: BaseStatisticFilter) {
    try {
      const rawExecutionTime =
        await this.tenderStatisticRepository.getTrackAverageTransaction(
          request,
        );

      const tracks = await this.tenderStatisticRepository.fetchAllTrack();

      const responseData: GetTrackAverageTransaction[] = [];

      if (rawExecutionTime.length > 0) {
        const lastRecord = rawExecutionTime[0].created_at;
        const groupedData = _.groupBy(
          rawExecutionTime,
          'proposal.project_track',
        );
        // console.log(logUtil(groupedData));

        for (const track of tracks) {
          const data = groupedData[track.id];
          let totalResponseTime = 0;
          let averageResponseTime = 0;
          let lastMonthTotalResponseTime = 0;
          let lastMonthAverageResponseTime = 0;

          if (data) {
            totalResponseTime = data.reduce(
              (sum, item) =>
                item.response_time !== null ? sum + item.response_time : sum,
              0,
            );
            averageResponseTime = totalResponseTime / data.length;

            // console.log(lastRecord);
            const lastMonth = moment(lastRecord).subtract(1, 'month');
            // console.log(lastMonth);
            const lastMonthData = data.filter((item) =>
              moment(item.created_at).isSame(lastMonth, 'month'),
            );

            // console.log(logUtil(lastMonthData));

            if (lastMonthData.length > 0) {
              lastMonthTotalResponseTime = lastMonthData.reduce(
                (sum, item) =>
                  item.response_time !== null ? sum + item.response_time : sum,
                0,
              );
              lastMonthAverageResponseTime =
                lastMonthTotalResponseTime / lastMonthData.length;
            }
          }

          responseData.push({
            project_track: track.id,
            raw_total_response_time: totalResponseTime,
            total_response_time:
              portalReportConvertMinutesToHours(totalResponseTime),
            raw_average_response_time: averageResponseTime,
            average_response_time:
              portalReportConvertMinutesToHours(averageResponseTime),
            raw_last_month_total_response_time: lastMonthTotalResponseTime,
            last_month_total_response_time: portalReportConvertMinutesToHours(
              lastMonthTotalResponseTime,
            ),
            raw_last_month_average_response_time: lastMonthAverageResponseTime,
            last_month_average_response_time: portalReportConvertMinutesToHours(
              lastMonthAverageResponseTime,
            ),
          });
        }
      } else {
        for (const track of tracks) {
          responseData.push({
            project_track: track.id,
            raw_total_response_time: 0,
            total_response_time: '0',
            raw_average_response_time: 0,
            average_response_time: '0',
            raw_last_month_total_response_time: 0,
            last_month_total_response_time: '0',
            raw_last_month_average_response_time: 0,
            last_month_average_response_time: '0',
          });
        }
      }

      return [...responseData];
    } catch (error) {
      throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
    }
  }

  async getEmployeeAverageTransaction(
    request: AverageEmployeeTransactionTimeFilter,
  ): Promise<GetAverageEmployeeResponseTime> {
    const response =
      await this.tenderStatisticRepository.getEmployeeAverageTransaction(
        request,
      );

    if (response.length > 0) {
      return {
        data: response.map((rawData: GetRawAverageEmployeeResponseTime) => {
          const data: AverageEmployeeResponseTime = {
            id: rawData.id.toString(),
            employee_name: rawData.employee_name,
            account_type: rawData.account_type,
            section: rawData.section,
            total_transaction: Number(rawData.total_transaction),
            raw_response_time: Number(rawData.response_time),
            response_time: portalReportConvertMinutesToHours(
              Number(rawData.response_time),
            ),
            raw_average_response_time: Number(rawData.average_response_time),
            average_response_time: portalReportConvertMinutesToHours(
              Number(rawData.average_response_time),
            ),
          };
          return data;
        }),
        total: Number(response[0].total),
      };
    }
    return {
      data: [],
      total: 0,
    };
  }
}
